cloud-functions/splitter-cleanup-service/requirements.txt
functions-framework==3.*
google-cloud-storage
moviepy==1.0.3
imageio-ffmpeg